server/python-backend/requirements.txt
flask
flask-cors
pandas
mysql-connector-python
python-dotenv
